Extract mutation tests and repair duplicated helper preamble
Address review feedback on the mutation-testing template tests.

- Move the mutation-testing constants, `_rendered_section`, the
  `_assert_mutation_*` helpers, and the three mutation tests out of
  `tests/test_template.py` into a dedicated
  `tests/test_mutation_testing_template.py`, dropping the imports that
  became unused. Both modules now sit within the 400 logical-line limit
  (277 and 300, from 577 combined).
- Remove a duplicated preamble in `tests/test_helpers.py`: a stray copy of
  the module docstring rendered as a no-op expression, plus repeated
  `if TYPE_CHECKING` and `REPO_ROOT` definitions. Only the authoritative
  definitions at the top of the file remain.
- Annotate `_TEST_ACTION_SHA` and `_DEPENDABOT_UPDATED_TEST_ACTION_SHA`
  as `str`.
- Add the missing comma before "so" in the parent developers' guide.
